+++ src/Interface/Results.php
@@ -33,12 +33,4 @@ interface Results extends \Countable
     public function fetchNth(int|string $n, ?string $dto = null): \Generator;
 
     public function exists(): bool;
-
-    public function sum(string $key): float;
-
-    public function avg(string $key, int $decimalPlaces = 2): float;
-
-    public function min(string $key): float;
-
-    public function max(string $key): float;
 }