+++ src/Traits/Groupable.php
@@ -2,15 +2,32 @@
 
 namespace FQL\Traits;
 
+use FQL\Exception;
+
 trait Groupable
 {
     /**
      * @var string[] $groupByFields
      */
     private array $groupByFields = [];
+    private bool $groupableBlocked = false;
+
+    public function blockGroupable(): void
+    {
+        $this->groupableBlocked = true;
+    }
+
+    public function isGroupableEmpty(): bool
+    {
+        return $this->groupByFields === [];
+    }
 
     public function groupBy(string ...$fields): self
     {
+        if ($this->groupableBlocked) {
+            throw new Exception\QueryLogicException('GROUP BY is not allowed in DESCRIBE mode');
+        }
+
         $this->groupByFields = array_merge($this->groupByFields, $fields);
         return $this;
     }